Activacion de sesiones en .env array

+++ app/Http/Controllers/FabricanteController.php
@@ -1,30 +1,50 @@
-<?php
-
-namespace App\Http\Controllers;
+<?php namespace App\Http\Controllers;
 
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
+
 use Illuminate\Http\Request;
-use Response;
-//cargamos fabricante porque lo usamos mas abajo
+
+// Cargamos Fabricante por que lo usamos más abajo.
 use App\Fabricante;
 
+use Response;
+
+// Activamos el uso de las funciones de caché.
+use Illuminate\Support\Facades\Cache;
+
 class FabricanteController extends Controller {
+
+	public function __construct()
+	{
+		$this->middleware('auth.basic',['only'=>['store','update','destroy']]);
+	}
 
 	/**
 	 * Display a listing of the resource.
 	 *
 	 * @return Response
 	 */
-	public function index() {
-		//return "en el index de fabricante
-		//Devolvemos un JSON cno todos los fabricantes
-		//return Fabricante::all();
-		//para devolver un json con codigo de respuesta HTTP
+	public function index()
+	{
+		// return "En el index de Fabricante.";
+		// Devolvemos un JSON con todos los fabricantes.
+		// return Fabricante::all();
 
-		return response()->json([
-					'status' => 'ok', 'data' => Fabricante::all()
-						], 200);
+		// Caché se actualizará con nuevos datos cada 15 segundos.
+		// cachefabricantes es la clave con la que se almacenarán 
+		// los registros obtenidos de Fabricante::all()
+		// El segundo parámetro son los minutos.
+		$fabricantes=Cache::remember('cachefabricantes',15/60,function()
+		{
+			return Fabricante::all();
+		});
+
+		// Para devolver un JSON con código de respuesta HTTP sin caché.
+		// return response()->json(['status'=>'ok', 'data'=>Fabricante::all()],200);
+
+		// Devolvemos el JSON usando caché.
+		return response()->json(['status'=>'ok', 'data'=>$fabricantes],200);
 	}
 
 	/**
@@ -32,36 +52,35 @@ class FabricanteController extends Controller {
 	 *
 	 * @return Response
 	 */
-	//No se utiliza este metodo porque se utiliza porque se usuaaria para mostrar un formulario de creacion de fabricantes y una api rest no hace eso 
-	public function create() {
+	// No se utiliza este método por que se usaría para mostrar un formulario
+	// de creación de Fabricantes. Y una API REST no hace eso.
+	/*
+	public function create()
+	{
 		//
 	}
+	*/
 
 	/**
 	 * Store a newly created resource in storage.
 	 *
 	 * @return Response
 	 */
-	public function store(Request $request) {
-		//
-		//metodo llamada al hacer un post
-		//comprobamos que recibimos todos los campos
-
-		if (!$request->input('nombre') || !$request->input('direccion') || !$request->input('telefono')) {
-			return response()->json([
-						'errors' => Array(['code' => 422, 'message' => 'Faltan datos necesarios para procesar el alta'
-							])], 422);
+	public function store(Request $request)
+	{
+		// Método llamado al hacer un POST.
+		// Comprobamos que recibimos todos los campos.
+		if (!$request->input('nombre') || !$request->input('direccion') || !$request->input('telefono'))
+		{
+			// NO estamos recibiendo los campos necesarios. Devolvemos error.
+			return response()->json(['errors'=>array(['code'=>422,'message'=>'Faltan datos necesarios para procesar el alta.'])],422);
 		}
 
-		//insertamos los datos recibidos en la tabla 
-		//devolvemos el codigo 201 que significa que se han insertados los datos
+		// Insertamos los datos recibidos en la tabla.
+		$nuevoFabricante=Fabricante::create($request->all());
 
-		$nuevoFabricante = Fabricante::create($request->all());
-
-		//devolvemos la respuesta Http 201 (created) + los datos del nuevo fabricante + una cabecera de Location
-
-		$respuesta = Response::make(json_encode(['data' => $nuevoFabricante]), 201)->header('Location', 'http://www.dominio.local/fabricante/' . $nuevoFabricante->id)->header('Content-Type', 'application/json');
-
+		// Devolvemos la respuesta Http 201 (Created) + los datos del nuevo fabricante + una cabecera de Location + cabecera JSON
+		$respuesta= Response::make(json_encode(['data'=>$nuevoFabricante]),201)->header('Location','http://www.dominio.local/fabricantes/'.$nuevoFabricante->id)->header('Content-Type','application/json');
 		return $respuesta;
 	}
 
@@ -71,19 +90,22 @@ class FabricanteController extends Controller {
 	 * @param  int  $id
 	 * @return Response
 	 */
-	public function show($id) {
-		//corresponde con la ruta /fabricantes/{fabbricante}
-		$fabricante = Fabricante::find($id);
+	public function show($id)
+	{
+		// Corresponde con la ruta /fabricantes/{fabricante}
+		// Buscamos un fabricante por el ID.
+		$fabricante=Fabricante::find($id);
 
-		if (!$fabricante) {
-			//se devuelve un error con los errores detectados y codigo 404
-
-			return response()->json([
-						'errors' => Array(['code' => 422, 'message' => 'No se encuentra un fabricante con ese codigo'
-							])], 422);
+		// Chequeamos si encontró o no el fabricante
+		if (! $fabricante)
+		{
+			// Se devuelve un array errors con los errores detectados y código 404
+			return response()->json(['errors'=>Array(['code'=>404,'message'=>'No se encuentra un fabricante con ese código.'])],404);
 		}
 
-		return response()->json(['status' => 'ok', 'data' => $fabricante], 200);
+		// Devolvemos la información encontrada.
+		return response()->json(['status'=>'ok','data'=>$fabricante],200);
+
 	}
 
 	/**
@@ -92,9 +114,12 @@ class FabricanteController extends Controller {
 	 * @param  int  $id
 	 * @return Response
 	 */
-	public function edit($id) {
+	/*
+	public function edit($id)
+	{
 		//
 	}
+	*/
 
 	/**
 	 * Update the specified resource in storage.
@@ -102,80 +127,83 @@ class FabricanteController extends Controller {
 	 * @param  int  $id
 	 * @return Response
 	 */
-	public function update($id, Request $request) {
-		//vamos a actualizar un fabricante
-		//comprobamos si el fabricante existe.  En otro caso devolvemos error
+	public function update($id,Request $request)
+	{
+		// Vamos a actualizar un fabricante.
+		// Comprobamos si el fabricante existe. En otro caso devolvemos error.
+		$fabricante=Fabricante::find($id);
 
-
-		$fabricante = Fabricante::find($id);
-
-		//si no existe mostrar error
-		if (!$fabricante) {
-			return response()->json([
-						'errors' => Array(['code' => 404, 'message' => 'No se encuentra un fabricante con ese codigo'
-							])], 404);
+		// Si no existe mostramos error.
+		if (! $fabricante)
+		{
+			// Devolvemos error 404.
+			return response()->json(['errors'=>array(['code'=>404,'message'=>'No se encuentra un fabricante con ese código.'])],404);
 		}
 
-		//almacenamos en variables para facilitar el uso, los campos recibidos
+		// Almacenamos en variables para facilitar el uso, los campos recibidos.
+		$nombre=$request->input('nombre');
+		$direccion=$request->input('direccion');
+		$telefono=$request->input('telefono');
 
-		$nombre = $request->input('nombre');
-		$direccion = $request->input('direccion');
-		$telefono = $request->input('telefono');
+		// Comprobamos si recibimos petición PATCH(parcial) o PUT (Total)
+		if ($request->method()=='PATCH')
+		{
+			$bandera=false;
 
-
-		if ($request->method() == 'PATCH') {
-			//actualizacion parcial de datos
-
-			$bandera = false;
-			if ($nombre != null && $nombre != '') {
-				$fabricante->nombre = $nombre;
-				$bandera = true;
+			// Actualización parcial de datos.
+			if ($nombre !=null && $nombre!='')
+			{
+				$fabricante->nombre=$nombre;
+				$bandera=true;
 			}
 
-
-			if ($direccion != null && $direccion != '') {
-				$fabricante->direccion = $direccion;
-				$bandera = true;
+			// Actualización parcial de datos.
+			if ($direccion !=null && $direccion!='')
+			{
+				$fabricante->direccion=$direccion;
+				$bandera=true;
 			}
 
-			if ($telefono != null && $telefono != '') {
-				$fabricante->$telefono = $telefono;
-				$bandera = true;
+			// Actualización parcial de datos.
+			if ($telefono !=null && $telefono!='')
+			{
+				$fabricante->telefono=$telefono;
+				$bandera=true;
 			}
 
-			if ($bandera) {
-				//devolvemos un codigo 200
+			if ($bandera)
+			{
+				// Grabamos el fabricante.
+				$fabricante->save();
 
-				return response()->json([
-							'status' => 'ok', 'data' => $fabricante]
-								, 200);
-			} else {
-
-				return response()->json([
-							'errors' => Array(['code' => 304, 'message' => 'No se ha modificado ningun dato'
-								])], 304);
-			}//else
-		}
-		
-		//metodo PUT actualizamos todos los campos
-		//comprobamos que recibimos todos
-		if (!$nombre || !$direccion || !$telefono) {
-			//se devuelve codigo 422 unprocessable Entity
-
-			return response()->json([
-						'errors' => Array(['code' => 422, 'message' => 'Faltan valores para completar precesamiento'
-							])], 422);
+				// Devolvemos un código 200.
+				return response()->json(['status'=>'ok','data'=>$fabricante],200);
+			}
+			else
+			{
+				// Devolvemos un código 304 Not Modified.
+				return response()->json(['errors'=>array(['code'=>304,'message'=>'No se ha modificado ningún dato del fabricante.'])],304);
+			}
 		}
 
 
-		//Actualizamos los 3 campos
-		$nombre = $request->input('nombre');
-		$direccion = $request->input('direccion');
-		$telefono = $request->input('telefono');
+		// Método PUT actualizamos todos los campos.
+		// Comprobamos que recibimos todos.
+		if (!$nombre || !$direccion || !$telefono)
+		{
+			// Se devuelve código 422 Unprocessable Entity.
+			return response()->json(['errors'=>array(['code'=>422,'message'=>'Faltan valores para completar el procesamiento.'])],422);
+		}
 
-		return response()->json([
-					'status' => 'ok', 'data' => $fabricante]
-						, 200);
+		// Actualizamos los 3 campos:
+		$fabricante->nombre=$nombre;
+		$fabricante->direccion=$direccion;
+		$fabricante->telefono=$telefono;
+
+		// Grabamos el fabricante
+		$fabricante->save();
+		return response()->json(['status'=>'ok','data'=>$fabricante],200);
+
 	}
 
 	/**
@@ -184,33 +212,43 @@ class FabricanteController extends Controller {
 	 * @param  int  $id
 	 * @return Response
 	 */
-	public function destroy($id) {
-		//
-		//borrado de fabricante
-		// fabricante/89 por DELETE
-		//comprobamos si el fabricante existe o no existe
+	public function destroy($id)
+	{
+		// Borrado de un fabricante.
+		// Ejemplo: /fabricantes/89 por DELETE
+		// Comprobamos si el fabricante existe o no.
+		$fabricante=Fabricante::find($id);
 
-		$fabricante = Fabricante::find($id);
+		if (! $fabricante)
+		{
+			// Devolvemos error codigo http 404
+			return response()->json(['errors'=>array(['code'=>404,'message'=>'No se encuentra el fabricante con ese código.'])],404);
+		}
 
-		//chequeamos si existe
+		// Borramos el fabricante y devolvemos código 204
+		// 204 significa "No Content".
+		// Este código no muestra texto en el body.
+		// Si quisiéramos ver el mensaje devolveríamos
+		// un código 200.
+		// Antes de borrarlo comprobamos si tiene aviones y si es así
+		// sacamos un mensaje de error.
+		// $aviones = $fabricante->aviones()->get();
+		$aviones = $fabricante->aviones;
 
-		if (!$fabricante) {
-			//devolvemos codigo http404
+		if (sizeof($aviones) >0)
+		{
+			// Si quisiéramos borrar todos los aviones del fabricante sería:
+			// $fabricante->aviones->delete();
 
-			return response()->json([
-						'errors' => Array(['code' => 404, 'message' => 'No se encuentra un fabricante con ese codigo'
-							])], 404);
-		}//if
+			// Devolvemos un código 409 Conflict. 
+			return response()->json(['errors'=>array(['code'=>409,'message'=>'Este fabricante posee aviones y no puede ser eliminado.'])],409);
+		}
 
-
+		// Eliminamos el fabricante si no tiene aviones.
 		$fabricante->delete();
 
-		return response()->json([
-					'code' => '204', 'message' => 'Se ha eliminado correctamente el fabricante'
-						], 204);
-
-		//borramos el fabricante y devolvemos codigo 204 (no content) sin contenido
-		//este codigo no muestra texto en el body
+		// Se devuelve código 204 No Content.
+		return response()->json(['code'=>204,'message'=>'Se ha eliminado correctamente el fabricante.'],204);
 	}
 
 }
